fix bug where encounters show as -1:00 long
I was able to track this down to an actual bad log statement in the raw
logs. This workaround should prevent this type of bug from happening.
Maybe the better thing to do is to limit which clocks we actually use
and then don't worry about it going back in time.